Move api-media-karma into api package (#688)
* Move api-media-karma into api package
* Make sure wepublish/api builds
* Adjsut KarmaMediaServer in example
