+++ package.json
@@ -11,9 +11,9 @@
     "dev": "run-s start:docker watch",
     "clean": "run-p --print-label clean:**",
     "test": "run-p --print-label test:**",
-    "build": "run-s --print-label build:api build:api-media-karma build:oauth2 build:editor build:api-example build:oauth2-example",
-    "build:production": "run-s --print-label build:api build:api-media-karma build:oauth2 build:editor:production",
-    "build:demo": "run-s --print-label build:api build:api-media-karma build:oauth2 build:editor:production build:api-example build:website-example build:oauth2-example",
+    "build": "run-s --print-label build:api build:oauth2 build:editor build:api-example build:oauth2-example",
+    "build:production": "run-s --print-label build:api build:oauth2 build:editor:production",
+    "build:demo": "run-s --print-label build:api build:oauth2 build:editor:production build:api-example build:website-example build:oauth2-example",
     "prerelease": "run-s clean build:production",
     "changelog": "node -r dotenv/config ./node_modules/.bin/lerna-changelog",
     "publish": "lerna publish from-package --ignore-scripts",
@@ -27,10 +27,6 @@
     "build:api": "yarn workspace @wepublish/api build",
     "watch:api": "yarn workspace @wepublish/api watch",
     "test:api": "yarn workspace @wepublish/api test",
-    "clean:api-media-karma": "yarn workspace @wepublish/api-media-karma clean",
-    "build:api-media-karma": "yarn workspace @wepublish/api-media-karma build",
-    "watch:api-media-karma": "yarn workspace @wepublish/api-media-karma watch",
-    "test:api-media-karma": "yarn workspace @wepublish/api-media-karma test",
     "setup:editor": "yarn workspace @wepublish/editor setup",
     "start:editor": "yarn workspace @wepublish/editor start",
     "clean:editor": "yarn workspace @wepublish/editor clean",
