--- packages/api-media-karma/.npmignore
@@ -1,1 +1,0 @@
-.tsbuildinfo
