--- packages/api-media-karma/CHANGELOG.md
@@ -1,16 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.8.5](https://github.com/wepublish/wepublish/compare/v0.8.4...v0.8.5) (2020-10-01)
-
-**Note:** Version bump only for package @wepublish/api-media-karma
-
-
-
-
-
-## [0.8.4](https://github.com/wepublish/wepublish/compare/v0.8.0...v0.8.4) (2020-08-12)
-
-**Note:** Version bump only for package @wepublish/api-media-karma
