--- packages/api-media-karma/__tests__/setup.ts
@@ -1,1 +1,0 @@
-/* file for setting up the environment before any test */
