--- packages/api-media-karma/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@wepublish/api-media-karma",
-  "version": "5.0.0",
-  "description": "We.publish media service",
-  "keywords": [
-    "images",
-    "postgres",
-    "publishing",
-    "media",
-    "newspaper",
-    "cms",
-    "subscription"
-  ],
-  "homepage": "https://github.com/wepublish/wepublish",
-  "bugs": {
-    "url": "https://github.com/wepublish/wepublish/issues",
-    "email": "development@wepublish.ch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/wepublish/wepublish.git",
-    "directory": "packages/api-media-karma"
-  },
-  "main": "./lib/",
-  "license": "MIT",
-  "scripts": {
-    "build": "run-s copy-license build:*",
-    "build:typescript": "tsc -b ./src/tsconfig.json",
-    "watch": "npm-run-all copy-license --parallel watch:*",
-    "watch:typescript": "tsc -b ./src/tsconfig.json --watch",
-    "copy-license": "shx cp ../../LICENSE ./",
-    "clean": "shx rm -rf ./LICENSE ./lib",
-    "prepare": "run-s clean copy-license build",
-    "test": "jest --coverage"
-  },
-  "dependencies": {
-    "form-data": "^3.0.0",
-    "node-fetch": "^2.6.0"
-  },
-  "devDependencies": {
-    "@types/graphql-upload": "^8.0.3",
-    "@types/jest": "^26.0.14",
-    "@types/node-fetch": "^2.5.5",
-    "jest": "^26.4.2",
-    "npm-run-all": "^4.1.5",
-    "shx": "^0.3.2",
-    "ts-jest": "^26.3.0",
-    "typescript": "^4.6.4"
-  },
-  "peerDependencies": {
-    "@wepublish/api": "*"
-  }
-}
