--- packages/api-media-karma/src/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "include": ["./**/*"],
-  "compilerOptions": {
-    "outDir": "../lib/",
-    "module": "commonjs",
-    "types": ["@types/node", "@types/jest"],
-    "lib": ["es2017"]
-  }
-}
