+++ packages/api/package.json
@@ -86,6 +86,7 @@
     "@types/ms": "^0.7.31",
     "@types/nanoid": "^2.0.0",
     "@types/node": "^13.7.7",
+    "@types/node-fetch": "^2.5.5",
     "@types/pino": "^6.3.5",
     "@types/pino-http": "^5.0.6",
     "@types/qs": "^6.9.1",
