+++ packages/api/src/context.ts
@@ -41,7 +41,7 @@ import {TokenExpiredError} from './error'
 import {Hooks} from './hooks'
 import {MailContext, MailContextOptions} from './mails/mailContext'
 import {BaseMailProvider} from './mails/mailProvider'
-import {MediaAdapter} from './mediaAdapter'
+import {MediaAdapter} from './media/mediaAdapter'
 import {MemberContext} from './memberContext'
 import {PaymentProvider} from './payments/paymentProvider'
 import {logger} from './server'
